moving general functions to super class (BaseImporter)
